 \name{CitationnbANDageAge}\docType{data}\alias{CitationnbANDageAge}\title{Fuzzy sets associated to each value of criteria Citation and reference Age, used in SCRIPTTEST}\description{This data set contains a table with specification of fuzzy sets for each value of Citation and reference Age. }\usage{CitationnbANDageAge}\format{An data frame containing fuzzy sets corresponding to Citation and reference Age values.}%\source{}%\references{%}\keyword{datasets}