 \name{Repetition}\docType{data}\alias{Repetition}\title{Fuzzy sets associated to each value of criteria Repetition, used in SCRIPTTEST}\description{This data set contains a table with specification of fuzzy sets for each value of Repetition. }\usage{Repetition}\format{An data frame containing fuzzy sets corresponding to Repetition values.}%\source{}%\references{%}\keyword{datasets}