\name{SCRIPTTEST}
\alias{SCRIPTTEST}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Test script}
\description{Test script for a method evaluating data reliability from meta-information.}
\usage{
SCRIPTTEST(del = "/")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
\item{del}{Used to access/write files}

}
%\details{
%}
\value{Return Final Evaluation}
%\references{
%% ~put references to the literature/web site here ~
%}
\author{N. Maillet, B. Charnomordic, S. Destercke}
%\note{
%%  ~~further notes~~
%}

\section{Warning }{....}
\examples{
SCRIPTTEST()
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line