\name{TypeOfSource}\docType{data}\alias{TypeOfSource}\title{Fuzzy sets associated to each value of criteria Type Of Source, used in SCRIPTTEST}\description{This data set contains a table with specification of fuzzy sets for each value of Type Of Source. }\usage{TypeOfSource}\format{An data frame containing fuzzy sets corresponding to Type Of Source values.}%\source{}%\references{%}\keyword{datasets}